requirements.txt
Flask==3.1.1
gunicorn==21.2.0
google-cloud-translate==3.12.1
